Initial project setup (#1)
* Clone and merge Hugo boilerplate Victor Hugo

This project will be built using the static site generator Hugo
(https://gohugo.io/). And Victor Hugo (https://github.com/netlify/victor-hugo)
is a boilerplate for setting up such a project. It will help us utilize modern
javascript (transpiled to legability code with Babel and Webpack) and css
(processed with PostCSS and CSSnext).

* Update Webpack and Gulp configs to match project

The Webpack config (WIP) that was supplied by the boilerplate didn't match the
needs of this project. Therefore it is updated with separate production and
development settings. Webpack is also updated from v1 to v2.

In the gulpfile only some code styling changes is made.

* Update development configs

This commit will update eslint to the needs of this project using airbnb's
preset.

Babel config is updated to instead use preset-env. This needs to be tweaked
further to match the needs of the project.

Flow is initialized and will aid during development.

* Add Yarn as dependency manager

And also add additional information about the project to package.json.

* Fix spelling in description

* Update README to reflect project